refactor: extract buildFormApi from useAbstractForm
Pulls the api-building block out of useAbstractForm into its own module
(src/runtime/core/build-form-api.ts). The new function is pure over
(FormState) → api with no dependency on the registry, consumer ref
counts, or the current Vue instance — those concerns stay in the
caller.

Enables phase 5.4's useFormContext to return a ReturnType-identical
API from an ambient injection without duplicating the wiring.
Behavioural no-op; all 507 tests pass unchanged.

Also registers kFormContext as a second injection key on the registry
module. useAbstractForm now calls provide(kFormContext, state) so
descendants can resolve the current form's state via
useFormContext() (added in the next commit). provide is safe during
setup regardless of whether useFormContext is ever called.
